Switched default architecure from amd64 to i386
Why? Becaus PC Vendors suck!
Hardware virtualization is deemed to 'dangerous' for
dumb PC users and is disabled in the BIOS on many PCs.
And as most PC users are to dumb to change an option in
their BIOS this .iso has to work without harware virtualization.
And Virtualbox only supports the half emulated virtualization mode
in x86 mode.

PC Vendors of the world: I hope you die in a fire.